app/requirements.txt
fastapi
uvicorn[standard]
selenium
webdriver-manager
python-dotenv
pydantic
google-genai
pymongo
